requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0
django-filter==24.1
djangorestframework-simplejwt==5.3.1
python-magic==0.4.27
django-cleanup==8.0.0
gunicorn==21.2.0
whitenoise==6.6.0 